refactor(kernel): split entity document schemas out of the kind registry
The registry crossed its 1100-line ceiling when the fact action catalog grew.
The explainable-schema helpers and the five lifecycle document schemas are
data with no dependency on the catalog or contract tables, so they move to
entity-document-schemas.ts as their own responsibility.
